==> 04_data_analysis/007_magma/22_pathway_analysis.qmd <==
---
title: Pathway analysis - MAGMA genes
execute:
  eval: true
#self-contained: true
editor_options: 
  chunk_output_type: console
---

```{r}
here::i_am("04_data_analysis/007_magma/22_pathway_analysis.qmd")
```

# Load packages

```{r}
source(here::here(
  "04_data_analysis/990_code_libraries/02_library-calls-for-renv.R"
))
```

# Read in data

Here we're looking at the significant celltypes from MAGMA again AD in all controls.
The genes are the genes MAGMA found to be significant after correction for each celltype.
Note that each celltype was adjusted separately and via Bonferroni.

```{r}
#| eval: true
# read data of sig magma genes from sig celltypes for AD all controls
res <- readr::read_tsv(here::here(
  "03_data/994_magma_inputs/results/sig_celltype_sig_genes.tsv"
)) |>
  dplyr::rename(celltype = SetName)
#res <- readr::read_tsv(here::here("03_data/994_magma_inputs/sig_genelist", "all_sig_magma_set_genes.tsv")) |>
#  dplyr::rename(celltype = SetName)
```

# Data prep

```{r}
# filter to sig genes
res_sig <- dplyr::filter(
  res,
  padj < 0.05 &
    input_gene_list == "all_controls" &
    percent == "top-ten-percent"
) |>
  dplyr::mutate(
    celltype = if_else(
      celltype == "Pericyte" &
        celltype_level == "level2",
      "Pericyte-1",
      celltype
    )
  ) |>
  dplyr::mutate(
    celltype = if_else(
      celltype == "Perivascular",
      "Perivascular-M",
      celltype
    )
  ) |>
  # remove gwas background not of interest
  dplyr::filter(grepl("EUROPEUKBB_35k10k", gwas_background)) |>
  dplyr::mutate(
    gwas_background = str_replace_all(
      gwas_background,
      "top-ten-percent",
      "case"
    ),
    case_control = if_else(grepl("_Case_", group), "case", "control")
  )

res_sig_level2 <- res_sig |>
  dplyr::filter(
    celltype_level == "level2" &
      gwas_background == "EUROPEUKBB_35k10k"
  )
readr::write_csv(
  res_sig_level2,
  "03_data/994_magma_inputs/sig_genelist/magma_level2_controls_sig_genes.csv"
)

# get gene names per celltype
genes_df <- res_sig |>
  dplyr::group_by(celltype, gwas_background, celltype_level) |>
  dplyr::summarise(list_of_columns = list(hgnc_symbol))
genes <- genes_df |>
  dplyr::pull(list_of_columns)
names(genes) <- paste0(
  genes_df$celltype,
  "-",
  genes_df$gwas_background,
  "-",
  genes_df$celltype_level
)
```

Let's check overlap between the case and control genes

```{r}
res_sig |>
  dplyr::select(celltype, celltype_level, hgnc_symbol, case_control)

df <- res_sig |>
  dplyr::select(celltype, celltype_level, hgnc_symbol, case_control) |>
  mutate(celltype_combined = paste(celltype, celltype_level, sep = "_"))

table(df$celltype, df$case_control)

df <- df |>
  dplyr::mutate(
    unique_genes_per_celltype = case_when(
      !duplicated(hgnc_symbol) & case_control == "case" ~ "case_only",
      !duplicated(hgnc_symbol) & case_control == "control" ~ "control_only",
      duplicated(hgnc_symbol) ~ "both",
      .default = NA
    ),
    .by = celltype_combined
  ) |>
  dplyr::mutate(
    unique_genes_overall = case_when(
      !duplicated(hgnc_symbol) & case_control == "case" ~ "case_only",
      !duplicated(hgnc_symbol) & case_control == "control" ~ "control_only",
      duplicated(hgnc_symbol) ~ "both",
      .default = NA
    )
  )

table(df$celltype_combined, df$unique_genes_per_celltype)
table(df$unique_genes_overall)

readr::write_tsv(
  df,
  here(
    "03_data/999_data_for_publication/case_risk/risk_genes_case_vs_control.tsv"
  )
)

table(df$celltype, df$unique_genes_per_celltype)

# Get the overall overlap in case/controls across all celltypes
binary_matrix <- res_sig |>
  dplyr::select(celltype, celltype_level, hgnc_symbol, case_control) |>
  mutate(celltype_combined = paste(celltype, celltype_level, sep = "_")) |>
  dplyr::select(!celltype:celltype_level) |>
  dplyr::distinct() |>
  dplyr::mutate(presence = 1) |>
  pivot_wider(
    names_from = case_control,
    values_from = presence,
    values_fill = list(presence = 0)
  )

# Prepare data for UpSet plot
upset_data <- binary_matrix |>
  select(-celltype_combined) |>
  group_by(hgnc_symbol) |>
  summarise_all(max) |>
  ungroup() |>
  column_to_rownames(var = "hgnc_symbol")

# Visualize using UpSet plot
library(UpSetR)
upset(upset_data, sets = c("case", "control"), keep.order = TRUE)
# Save plot
ggsave(here("03_data/999_data_for_publication/case_risk/upset_overall.png"))

unique_genes <- upset_data |>
  dplyr::filter(case == 1 & control == 0 | case == 0 & control == 1) |>
  rownames_to_column("hgnc_symbol") |>
  dplyr::left_join(df, by = join_by(hgnc_symbol))


# Step 1: Filter out celltypes not present in both case and control groups
filtered_df <- df |>
  dplyr::filter(n() > 0, .by = c(celltype_combined, case_control)) |>
  dplyr::filter(n_distinct(case_control) == 2, .by = celltype_combined)

# Step 2: Create a binary matrix indicating presence of genes in case/control groups for each celltype_combined
binary_matrix <- filtered_df |>
  mutate(presence = 1) |>
  pivot_wider(
    names_from = case_control,
    values_from = presence,
    values_fill = list(presence = 0)
  )

# Step 3: Prepare data for UpSet plot
upset_data <- binary_matrix |>
  dplyr::select(-celltype_combined) |>
  dplyr::summarise(across(where(is.numeric), max), .by = hgnc_symbol) |>
  column_to_rownames(var = "hgnc_symbol")

# Step 4: Visualize using UpSet plot
upset(upset_data, sets = c("case", "control"), keep.order = TRUE)
ggsave(here("03_data/999_data_for_publication/case_risk/upset_microglia-a.png"))
```

# Hypergeometric enrichment

```{r}
go_results_bp <- map(
  genes,
  ~ enrichGO(gene = .x, OrgDb = "org.Hs.eg.db", keyType = "SYMBOL", ont = "BP")
)
names(go_results_bp) <- paste0(names(go_results_bp), "_bp")
go_results_cc <- map(
  genes,
  ~ enrichGO(gene = .x, OrgDb = "org.Hs.eg.db", keyType = "SYMBOL", ont = "CC")
)
names(go_results_cc) <- paste0(names(go_results_cc), "_cc")
go_results_mf <- map(
  genes,
  ~ enrichGO(gene = .x, OrgDb = "org.Hs.eg.db", keyType = "SYMBOL", ont = "MF")
)
names(go_results_mf) <- paste0(names(go_results_mf), "_mf")
go_results <- c(go_results_bp, go_results_cc, go_results_mf)
rm(go_results_bp, go_results_cc, go_results_mf)
```

```{r}
# exclude cases where no pathways were found
go_results_filtered <- go_results |>
  purrr::keep(~ !is.null(.)) |>
  purrr::keep(~ nrow(.) > 0)

# treeplots
go_results_filtered <- map(go_results_filtered, pairwise_termsim)

# Define a safe version of your treeplot function that returns NULL on error
safe_treeplot <- possibly(treeplot, otherwise = NULL)

plan("multisession", workers = parallel::detectCores() - 1)
treeplots <- future_map(go_results_filtered, safe_treeplot)
# Remove any NULL entries
treeplots <- treeplots[!sapply(treeplots, is.null)]
# Assert that there are no nulls in list
assertthat::see_if(sum(map_lgl(treeplots, is.null)) == 0)
# Replace those with no treeplot with dotplots
no_treeplot <- names(go_results_filtered)[
  !names(go_results_filtered) %in%
    names(treeplots)
]
dotplots <- map(go_results_filtered[no_treeplot], dotplot)
assertthat::see_if(sum(map_lgl(dotplots, is.null)) == 0)
treeplots <- c(treeplots, dotplots)

treeplots <- map2(treeplots, names(treeplots), ~ .x + ggtitle(.y))

# Define the function to subset a single string up to the last underscore
subset_until_last_underscore <- function(input_string) {
  # Find the position of the last underscore
  last_underscore_position <- str_locate_all(input_string, "_")[[1]][, 1] |>
    max()

  # Subset the string up to the last underscore
  if (!is.na(last_underscore_position)) {
    return(str_sub(input_string, 1, last_underscore_position - 1))
  } else {
    return(input_string)
  }
}

# Apply the function to each element of the vector
celltypes <- sapply(names(treeplots), subset_until_last_underscore) |>
  unique()


merge_plots <- function(celltype, dotplot_list) {
  # subset to relevant plots
  plots <- dotplot_list[grepl(celltype, names(dotplot_list))]
  # Merge plots
  plot <- plots[[1]] + plots[[2]] + plots[[3]]
}

plots <- map(celltypes, merge_plots, treeplots) |> set_names(celltypes)

plots

# Save plots
map2(
  plots,
  names(plots),
  ~ ggsave(
    here::here(
      "05_figures/990_shared_figures/001_magma/magma_pathways",
      paste0(.y, "_sig_genes_only", ".png")
    ),
    plot = .x,
    width = 32,
    height = 10
  )
)

dotplots <-
  map2(
    go_results_filtered,
    names(go_results_filtered),
    ~ dotplot(.x, showCategory = 20) +
      ggtitle(.y)
  )

plots <- map(celltypes, merge_plots, dotplots) |> set_names(celltypes)

map2(
  plots,
  names(plots),
  ~ ggsave(
    here::here(
      "05_figures/990_shared_figures/001_magma/magma_pathways",
      paste0(.y, "_sig_genes_only-dotplots", ".png")
    ),
    plot = .x,
    width = 26,
    height = 12
  )
)
```

```{r}
celltypes <- sub("_.*", "", names(go_results_filtered)) |> unique()

merged_go_plot <- function(celltype, go_results_list) {
  # Subset to celltype
  gos <- go_results_list[grepl(celltype, names(go_results_list))]

  # Extract results and add ontology information
  combined_df <- do.call(
    rbind,
    lapply(names(gos), function(name) {
      df <- as.data.frame(gos[[name]]@result)
      df$Ontology <- gsub(".*_(.*)", "\\1", name) # Extract ontology from name
      # get the celltype level from name
      df$celltype_level <- str_extract(name, "(?<=-)[^-]*$") |>
        str_sub(end = -4)

      return(df)
    })
  )
  return(combined_df)
}

go_combined <- map(celltypes, merged_go_plot, go_results_filtered) |>
  set_names(celltypes)
# Save
qs::qsave(
  go_combined,
  here::here(
    "05_figures/990_shared_figures/003_final_figures/001_data_for_plots/go_data_sig_gene_only.qs"
  )
)
```

# Gene set enrichment analysis (GSEA)

- Note that GSEA takes pval and foldchange into account when computing the enrichment, so we don't filter the gene list

```{r}
# nest data by celltype
gsea_df <- res |>
  dplyr::filter(
    input_gene_list == "all_controls" &
      percent == "top-ten-percent"
  ) |>
  dplyr::mutate(
    celltype = if_else(
      celltype == "Pericyte" &
        celltype_level == "level2",
      "Pericyte-1",
      celltype
    )
  ) |>
  # remove gwas background not of interest
  dplyr::filter(grepl("EUROPEUKBB_35k10k", gwas_background)) |>
  dplyr::mutate(
    gwas_background = str_replace_all(
      gwas_background,
      "top-ten-percent",
      "case"
    )
  ) |>
  # group by celltype
  dplyr::group_by(celltype, gwas_background, celltype_level) |>
  tidyr::nest()
```


```{r}
#| message: false
#| warning: false
# Function: Adjacency matrix to list -------------------------
matrix_to_list <- function(pws) {
  pws.l <- list()
  for (pw in colnames(pws)) {
    pws.l[[pw]] <- rownames(pws)[as.logical(pws[, pw])]
  }
  return(pws.l)
}

## Function: prepare_gmt --------------------------------------
prepare_gmt <- function(gmt_file, genes_in_data, savefile = FALSE) {
  # for debug
  #file <- gmt_files[1]
  #genes_in_data <- df$gene_symbol

  # Read in gmt file
  gmt <- gmtPathways(gmt_file)
  hidden <- unique(unlist(gmt))

  # Convert gmt file to a matrix with the genes as rows and for each go annotation (columns) the values are 0 or 1
  mat <- matrix(
    NA,
    dimnames = list(hidden, names(gmt)),
    nrow = length(hidden),
    ncol = length(gmt)
  )
  for (i in 1:dim(mat)[2]) {
    mat[, i] <- as.numeric(hidden %in% gmt[[i]])
  }

  #Subset to the genes that are present in our data to avoid bias
  hidden1 <- intersect(genes_in_data, hidden)
  mat <-
    mat[hidden1, colnames(mat)[which(colSums(mat[hidden1, ]) > 5)]] # filter for gene sets with more than 5 genes annotated
  # And get the list again
  final_list <-
    matrix_to_list(mat) # for this we use the function we previously defined

  if (savefile) {
    saveRDS(
      final_list,
      file = paste0(
        gsub('.gmt', '', gmt_file),
        '_subset_',
        format(Sys.time(), '%d%m'),
        '.RData'
      )
    )
  }

  print('Wohoo! .gmt conversion successfull!:)')
  return(final_list)
}

# Get all the genes in dataset and assign them to my_genes
my_genes <- res$hgnc_symbol

# Download gene sets .gmt files
#https://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp
# I've downloaded several background sets to try
gmt_files <- list.files(
  here::here("03_data/991_external_data"),
  pattern = "\\.gmt$",
  full.names = TRUE
)

gmt <- map(gmt_files, ~ prepare_gmt(.x, my_genes))
```

```{r}
#| eval: false
#| include: false

# This is the per celltype version - not using fgsea at the moment so I'll skip
# this for now
# Get all the genes in your dataset and assign them to my_genes
my_genes <- map(gsea_df$data, ~ .x$gene) |>
  purrr::set_names(gsea_df$celltype)

gmt <- map(gmt_files, function(gmt) {
  map(my_genes, function(genes) {
    prepare_gmt(gmt, genes)
  })
})
```

Get the gene rankings

```{r}
# Genes need to be ranked for GSEA as named gene list
# Here I'll use use the signed p values from spatial DGE as ranking

# Function to sort rankings
sort_rankings <- function(log2fc, pvals, gene_names, group, plot = FALSE) {
  rankings <- sign(log2fc) * (-log10(pvals))
  # genes as names
  names(rankings) <- gene_names
  head(rankings)

  # sort genes by ranking
  rankings <- sort(rankings, decreasing = TRUE)
  if (plot) {
    plot(rankings, main = group)
  }

  # Some genes have such low p values that the signed pval is +- inf, we need to
  # change it to the maximum * constant to avoid problems with fgsea
  max_ranking <- max(rankings[is.finite(rankings)])
  min_ranking <- min(rankings[is.finite(rankings)])
  # change any inf values to 10 times the min/max non-inf value
  rankings <-
    replace(rankings, rankings > max_ranking, max_ranking * 10)
  rankings <-
    replace(rankings, rankings < min_ranking, min_ranking * 10)
  rankings <-
    sort(rankings, decreasing = TRUE) # sort genes by ranking

  return(rankings)
}
```

```{r}
#| eval: false
#| include: false

# Again for more complex contrasts
rankings <-
  map2(
    gsea_df$data,
    gsea_df$celltype_contrast,
    ~ (sort_rankings(.x$log2FoldChange, .x$pvalue, .x$gene, .y))
  ) |>
  purrr::set_names(gsea_df$celltype_contrast)
```

```{r}
rankings <-
  map2(
    gsea_df$data,
    gsea_df$celltype,
    ~ (sort_rankings(.x$ZSTAT, .x$P, .x$hgnc_symbol, .y))
  ) |>
  purrr::set_names(paste0(
    gsea_df$celltype,
    "-",
    gsea_df$gwas_background,
    "-",
    gsea_df$celltype_level
  ))
```



```{r}
ggplot(
  data.frame(
    gene_symbol = names(
      rankings$`Microglia-activated-case_EUROPEUKBB_35k10k-level2`
    )[1:50],
    ranks = rankings$`Microglia-activated-case_EUROPEUKBB_35k10k-level2`[1:50]
  ),
  aes(gene_symbol, ranks)
) +
  geom_point() +
  theme_classic() +
  theme(
    axis.text.x = element_text(
      angle = 90,
      vjust = 0.5,
      hjust = 1
    )
  ) +
  ggtitle(
    "Microglia-activated top 50 ranked genes",
    subtitle = "MAGMA sig genes"
  )
```

## Run GSEA

```{r}
#| message: false
#| results: hide
## Run GSEA ---------------------------------------------------------------
gsea_res <- map(gmt, function(pathway) {
  map(rankings, function(rank) {
    fgsea(
      pathways = pathway,
      # List of gene sets to check
      stats = rank,
      # in this case we have both pos and neg rankings. if only pos or neg, set
      # to 'pos', 'neg'
      scoreType = 'std',
      minSize = 10,
      maxSize = 500,
      # for parallelisation
      nproc = 8
    )
  })
}) |>
  purrr::set_names(basename(gmt_files))
```


```{r}
# get log2 FC as a vector with the respective gene names, sorted in decreasing
# order
named_vectors <- res_sig |>
  dplyr::group_by(celltype, gwas_background, celltype_level) |>
  tidyr::nest() |>
  dplyr::mutate(
    gene_list = map(
      data,
      ~ setNames(.x$ZSTAT, .x$hgnc_symbol) |>
        sort(decreasing = TRUE)
    )
  )
```

## Gene Ontology

```{r}
# Define the file path to save the resulting object
result_file <-
  here::here(
    "03_data/990_processed_data/007_magma/001_pathway_analysis/gene-set-enrichment_go_magma.rds"
  )

# Check if the file exists
if (!file.exists(result_file)) {
  plan(multisession, workers = parallel::detectCores() - 1)
  # Code for your time-consuming analysis
  gse_bp <-
    future_map(
      rankings,
      ~ gseGO(
        .x,
        OrgDb = "org.Hs.eg.db",
        ont = "BP",
        keyType = "SYMBOL"
      ),
      .options = furrr_options(seed = 123)
    )
  gse_cc <-
    future_map(
      rankings,
      ~ gseGO(
        .x,
        OrgDb = "org.Hs.eg.db",
        ont = "CC",
        keyType = "SYMBOL"
      ),
      .options = furrr_options(seed = 123)
    )
  gse_mf <-
    future_map(
      rankings,
      ~ gseGO(
        .x,
        OrgDb = "org.Hs.eg.db",
        ont = "MF",
        keyType = "SYMBOL"
      ),
      .options = furrr_options(seed = 123)
    )

  names(gse_bp) <- paste0(names(gse_bp), "_BP")
  names(gse_mf) <- paste0(names(gse_mf), "_MF")
  names(gse_cc) <- paste0(names(gse_cc), "_CC")
  gse <- c(gse_bp, gse_mf, gse_cc)
  rm(gse_bp, gse_mf, gse_cc)

  # Save the resulting object
  readr::write_rds(gse, result_file)
} else {
  # Load the existing resulting object
  gse <- readr::read_rds(result_file)
}
```

```{r}
# remove any with no pathways found
gse <- gse |>
  purrr::keep(~ nrow(.) > 0)
```

### Insulin pathways

Zam is interested in pathways related to insulin, so I'll filter to celltypes where such terms are present

```{r}
# Function to check for the presence of 'insulin' in the pathway description
find_insulin <- function(df) {
  df <- as_tibble(df)
  grepl("insulin", df$Description, ignore.case = TRUE)
}

# Apply the function to each dataframe in the list
insulin_related <- lapply(gse, find_insulin)

# Now, determine which celltypes have any insulin-related pathways
celltypes_with_insulin <- names(insulin_related)[sapply(insulin_related, any)]

# Print the celltypes with insulin-related pathways
print(celltypes_with_insulin)

# Remove ambiguous
celltypes_with_insulin <-
  celltypes_with_insulin[!celltypes_with_insulin == "ambiguous"]

print("Celltype with 'insulin' in pathway description:")
paste0(celltypes_with_insulin, collapse = ", ")
```

Get the genes of these celltypes too

```{r}
#| eval: false

# Save the dotplots of this subset
insulin_pathways <- gse[names(gse) %in% celltypes_with_insulin]
dotplots <-
  map2(
    insulin_pathways,
    names(insulin_pathways),
    ~ dotplot(.x, showCategory = 10, split = ".sign") +
      facet_grid(. ~ .sign) +
      ggtitle(.y)
  )

# Specify the filename for the PDF
if (include_apoe) {
  pdf_file <- here::here("insulin_celltype_dotplots_apoe.pdf")
} else {
  pdf_file <- here::here("insulin_celltype_dotplots_no-apoe.pdf")
}
# Open the PDF device
pdf(pdf_file)
dotplots
# Close the PDF device
dev.off()

insulin_genes <- res_sig |>
  dplyr::filter(celltype %in% celltypes_with_insulin)
insulin_genes |> dplyr::filter(celltype == "Capillary") |> dim()
if (include_apoe) {
  readr::write_csv(
    insulin_genes,
    here::here("insulin_celltypes_sig_genes_apoe.csv")
  )
} else {
  readr::write_csv(
    insulin_genes,
    here::here("insulin_celltypes_sig_genes_no-apoe.csv")
  )
}
```

### Plots

```{r}
dotplots <-
  map2(
    gse,
    names(gse),
    ~ dotplot(.x, showCategory = 10, split = ".sign") +
      facet_grid(. ~ .sign) +
      ggtitle(.y)
  )

# Use the sub function to extract the substring before the first underscore
celltypes <- sub("_.*", "", names(dotplots)) |> unique()

merge_plots <- function(celltype, dotplot_list) {
  # subset to relevant plots
  plots <- dotplot_list[grepl(celltype, names(dotplot_list))]
  # Merge plots
  plot <- plots[[1]] + plots[[2]] + plots[[3]]
}

plots <- map(celltypes, merge_plots, dotplots) |> set_names(celltypes)

plots

# Save plots
map2(
  plots,
  names(plots),
  ~ ggsave(
    here::here(
      "05_figures/990_shared_figures/001_magma/magma_pathways",
      paste0(.y, ".png")
    ),
    plot = .x,
    width = 18
  )
)
```

```{r}
# treeplots
gse_tree <- map(gse, pairwise_termsim)

# Define a safe version of your treeplot function that returns NULL on error
safe_treeplot <- possibly(treeplot, otherwise = NULL)

plan("multisession", workers = parallel::detectCores() - 1)
treeplots <- future_map(gse_tree, safe_treeplot)
# Remove any NULL entries
treeplots <- treeplots[!sapply(treeplots, is.null)]
# Assert that there are no nulls in list
assertthat::see_if(sum(map_lgl(treeplots, is.null)) == 0)
# Replace those with no treeplot with dotplots
no_treeplot <- names(gse_tree)[
  !names(gse_tree) %in%
    names(treeplots)
]
dotplots <- map(gse_tree[no_treeplot], dotplot)
assertthat::see_if(sum(map_lgl(dotplots, is.null)) == 0)
treeplots <- c(treeplots, dotplots)

treeplots <- map2(treeplots, names(treeplots), ~ .x + ggtitle(.y))

plots <- map(celltypes, merge_plots, treeplots) |> set_names(celltypes)

map2(
  plots,
  names(plots),
  ~ ggsave(
    here::here(
      "05_figures/990_shared_figures/001_magma/magma_pathways",
      paste0("gsea_treeplots_", .y, ".png")
    ),
    plot = .x,
    width = 32,
    height = 10
  )
)
```



```{r}
#| eval: false
#| include: false
map2(
  gse,
  names(gse),
  ~ emapplot(.x, showCategory = 10) +
    ggtitle(.y)
)
```

```{r}
map2(
  gse,
  names(gse),
  ~ ridgeplot(.x) +
    labs(x = "enrichment distribution") +
    ggtitle(.y)
)
```

Note, one can change which pathway is plotted here by setting the title and `geneSetID` accordingly

```{r}
map2(
  gse,
  names(gse),
  ~ gseaplot(.x, by = "all", title = .x$Description[1], geneSetID = 1)
)
```
